lib: Widen iommu_config page_size_mask to 64 bit
The virtio-iommu config page_size_mask is a 64 bit field. The
truncated view declared it 32 bit, which shifted input_range and
domain_range by four bytes. On AArch64 the misaligned 64 bit read
of input_range faulted the guest. Match the spec layout so I0094
and I0095 read the correct aligned fields.

+++ lib/virtio_iommu.h
@@ -100,7 +100,7 @@ struct virtio_iommu_req_probe {
 
 /* Truncated device config view used by capability boundary tests. */
 struct iommu_config {
-    uint32_t page_size_mask;
+    uint64_t page_size_mask;
     struct { uint64_t start, end; } input_range;
     struct { uint32_t start, end; } domain_range;
 } __attribute__((packed));
